Use global mutex instead critical section so that multiple instance don't init at once in case IIS starts multiple processes
git-svn-id: https://svn.apache.org/repos/asf/tomcat/jk/trunk@1297439 13f79535-47bb-0310-9956-ffa450edef68

+++ native/common/jk_uri_worker_map.c
@@ -1075,7 +1075,7 @@ const char *map_uri_to_worker_ext(jk_uri_worker_map_t *uw_map,
             off = 1;
         }
         /* Size including leading slash. */
-        vhost_len = strlen(vhost);
+        vhost_len = (unsigned int)strlen(vhost);
         if (vhost_len + off >= JK_MAX_URI_LEN) {
             vhost_len = 0;
             jk_log(l, JK_LOG_WARNING,
